+++ fltk/src/group.rs
@@ -339,22 +339,24 @@ impl Pack {
     }
 }
 
-/// Defines a Vertical Grid (custom widget).
-/// Requires setting the params manually using the `set_params` method, which takes the rows, columns and spacing.
-/// Requires explicit calls to add, which is overloaded especially for the layout.
-/// ```rust,no_run
-/// use fltk::{prelude::*, *};
-/// let mut grid = group::VGrid::new(0, 0, 400, 300, "");
-/// grid.set_params(3, 3, 5);
-/// grid.add(&button::Button::default());
-/// grid.add(&button::Button::default());
-/// grid.add(&button::Button::default());
-/// grid.add(&button::Button::default());
-/// grid.add(&button::Button::default());
-/// grid.add(&button::Button::default());
-/// grid.add(&button::Button::default());
-/// grid.add(&button::Button::default());
-/// ```
+/**
+    Defines a Vertical Grid (custom widget).
+    Requires setting the params manually using the `set_params` method, which takes the rows, columns and spacing.
+    Requires explicit calls to add, which is overloaded especially for the layout.
+    ```rust,no_run
+    use fltk::{prelude::*, *};
+    let mut grid = group::VGrid::new(0, 0, 400, 300, "");
+    grid.set_params(3, 3, 5);
+    grid.add(&button::Button::default());
+    grid.add(&button::Button::default());
+    grid.add(&button::Button::default());
+    grid.add(&button::Button::default());
+    grid.add(&button::Button::default());
+    grid.add(&button::Button::default());
+    grid.add(&button::Button::default());
+    grid.add(&button::Button::default());
+    ```
+*/
 #[derive(Debug, Clone)]
 pub struct VGrid {
     vpack: Pack,
@@ -421,22 +423,24 @@ impl DerefMut for VGrid {
     }
 }
 
-/// Defines a Horizontal Grid (custom widget).
-/// Requires setting the params manually using the `set_params` method, which takes the rows, columns and spacing.
-/// Requires explicit calls to add, which is overloaded especially for the layout.
-/// ```rust,no_run
-/// use fltk::{prelude::*, *};
-/// let mut grid = group::HGrid::new(0, 0, 400, 300, "");
-/// grid.set_params(3, 3, 5);
-/// grid.add(&button::Button::default());
-/// grid.add(&button::Button::default());
-/// grid.add(&button::Button::default());
-/// grid.add(&button::Button::default());
-/// grid.add(&button::Button::default());
-/// grid.add(&button::Button::default());
-/// grid.add(&button::Button::default());
-/// grid.add(&button::Button::default());
-/// ```
+/**
+    Defines a Horizontal Grid (custom widget).
+    Requires setting the params manually using the `set_params` method, which takes the rows, columns and spacing.
+    Requires explicit calls to add, which is overloaded especially for the layout.
+    ```rust,no_run
+    use fltk::{prelude::*, *};
+    let mut grid = group::HGrid::new(0, 0, 400, 300, "");
+    grid.set_params(3, 3, 5);
+    grid.add(&button::Button::default());
+    grid.add(&button::Button::default());
+    grid.add(&button::Button::default());
+    grid.add(&button::Button::default());
+    grid.add(&button::Button::default());
+    grid.add(&button::Button::default());
+    grid.add(&button::Button::default());
+    grid.add(&button::Button::default());
+    ```
+*/
 #[derive(Debug, Clone)]
 pub struct HGrid {
     hpack: Pack,