+++ fltk/src/lib.rs
@@ -47,7 +47,7 @@ For faster builds you can enable ninja builds for the C++ source using the "use-
 
 An example hello world application:
 
-```no_run
+```rust,no_run
 use fltk::{app, prelude::*, window::Window};
 let app = app::App::default();
 let mut wind = Window::new(100, 100, 400, 300, "Hello from rust");
@@ -57,7 +57,7 @@ app.run().unwrap();
 ```
 
 Another example showing the basic callback functionality:
-```no_run
+```rust,no_run
 use fltk::{app, button::Button, frame::Frame, prelude::*, window::Window};
 let app = app::App::default();
 let mut wind = Window::new(100, 100, 400, 300, "Hello from rust");
@@ -71,7 +71,7 @@ app.run().unwrap();
 Please check the examples directory for more examples.
 You will notice that all widgets are instantiated with a new() method, taking the x and y coordinates, the width and height of the widget, as well as a label which can be left blank if needed. Another way to initialize a widget is using the builder pattern: (The following buttons are equivalent)
 
-```no_run
+```rust,no_run
 use fltk::{button::*, prelude::*};
 let but1 = Button::new(10, 10, 80, 40, "Button 1");
 
@@ -82,7 +82,7 @@ let but2 = Button::default()
 ```
 
 An example of a counter showing use of the builder pattern:
-```no_run
+```rust,no_run
 use fltk::{app, button::*, frame::*, prelude::*, window::*};
 let app = app::App::default();
 let mut wind = Window::default()
@@ -107,7 +107,7 @@ wind.show();
 /* Event handling */
 ```
 Alternatively, you can use packs to layout your widgets:
-```no_run
+```rust,no_run
 use fltk::{button::*, frame::*, group::*, prelude::*, window::*};
 let mut wind = Window::default().with_size(160, 200).with_label("Counter");
 // Vertical is default. You can choose horizontal using pack.set_type(PackType::Horizontal);